fix: optimize Cycle size to improve tracer memory usage

+++ common/src/constants.rs
@@ -1,7 +1,7 @@
 pub const XLEN: usize = 32;
-const RISCV_REGISTER_COUNT: u64 = 32;
-const VIRTUAL_REGISTER_COUNT: u64 = 32; //  see Section 6.1 of Jolt paper
-pub const REGISTER_COUNT: u64 = RISCV_REGISTER_COUNT + VIRTUAL_REGISTER_COUNT;
+const RISCV_REGISTER_COUNT: u8 = 32;
+const VIRTUAL_REGISTER_COUNT: u8 = 32; //  see Section 6.1 of Jolt paper
+pub const REGISTER_COUNT: u8 = RISCV_REGISTER_COUNT + VIRTUAL_REGISTER_COUNT;
 pub const BYTES_PER_INSTRUCTION: usize = 4;
 
 pub const RAM_START_ADDRESS: u64 = 0x80000000;
@@ -18,7 +18,7 @@ pub const DEFAULT_MAX_INPUT_SIZE: u64 = 4096;
 pub const DEFAULT_MAX_OUTPUT_SIZE: u64 = 4096;
 pub const DEFAULT_MAX_TRACE_LENGTH: u64 = 1 << 24;
 
-pub const fn virtual_register_index(index: u64) -> u64 {
+pub const fn virtual_register_index(index: u8) -> u8 {
     index + VIRTUAL_REGISTER_COUNT
 }
