+++ tracer/src/instruction/div.rs
@@ -31,14 +31,14 @@ declare_riscv_instr!(
 
 impl DIV {
     fn exec(&self, cpu: &mut Cpu, _: &mut <DIV as RISCVInstruction>::RAMAccess) {
-        let dividend = cpu.x[self.operands.rs1];
-        let divisor = cpu.x[self.operands.rs2];
+        let dividend = cpu.x[self.operands.rs1 as usize];
+        let divisor = cpu.x[self.operands.rs2 as usize];
         if divisor == 0 {
-            cpu.x[self.operands.rd] = -1;
+            cpu.x[self.operands.rd as usize] = -1;
         } else if dividend == cpu.most_negative() && divisor == -1 {
-            cpu.x[self.operands.rd] = dividend;
+            cpu.x[self.operands.rd as usize] = dividend;
         } else {
-            cpu.x[self.operands.rd] = cpu.sign_extend(dividend.wrapping_div(divisor))
+            cpu.x[self.operands.rd as usize] = cpu.sign_extend(dividend.wrapping_div(divisor))
         }
     }
 }
@@ -47,8 +47,8 @@ impl RISCVTrace for DIV {
     fn trace(&self, cpu: &mut Cpu, trace: Option<&mut Vec<RV32IMCycle>>) {
         // RISCV spec: For REM, the sign of a nonzero result equals the sign of the dividend.
         // DIV operands
-        let x = cpu.x[self.operands.rs1];
-        let y = cpu.x[self.operands.rs2];
+        let x = cpu.x[self.operands.rs1 as usize];
+        let y = cpu.x[self.operands.rs2 as usize];
 
         let (quotient, remainder) = match cpu.xlen {
             Xlen::Bit32 => {
@@ -98,10 +98,10 @@ impl RISCVTrace for DIV {
 impl VirtualInstructionSequence for DIV {
     fn virtual_sequence(&self) -> Vec<RV32IMInstruction> {
         // Virtual registers used in sequence
-        let v_0 = virtual_register_index(0) as usize;
-        let v_q = virtual_register_index(1) as usize;
-        let v_r = virtual_register_index(2) as usize;
-        let v_qy = virtual_register_index(3) as usize;
+        let v_0 = virtual_register_index(0);
+        let v_q = virtual_register_index(1);
+        let v_r = virtual_register_index(2);
+        let v_qy = virtual_register_index(3);
 
         let mut sequence = vec![];
