+++ tracer/src/instruction/rem.rs
@@ -30,15 +30,16 @@ declare_riscv_instr!(
 
 impl REM {
     fn exec(&self, cpu: &mut Cpu, _: &mut <REM as RISCVInstruction>::RAMAccess) {
-        let dividend = cpu.x[self.operands.rs1];
-        let divisor = cpu.x[self.operands.rs2];
+        let dividend = cpu.x[self.operands.rs1 as usize];
+        let divisor = cpu.x[self.operands.rs2 as usize];
         if divisor == 0 {
-            cpu.x[self.operands.rd] = dividend;
+            cpu.x[self.operands.rd as usize] = dividend;
         } else if dividend == cpu.most_negative() && divisor == -1 {
-            cpu.x[self.operands.rd] = 0;
+            cpu.x[self.operands.rd as usize] = 0;
         } else {
-            cpu.x[self.operands.rd] =
-                cpu.sign_extend(cpu.x[self.operands.rs1].wrapping_rem(cpu.x[self.operands.rs2]));
+            cpu.x[self.operands.rd as usize] = cpu.sign_extend(
+                cpu.x[self.operands.rs1 as usize].wrapping_rem(cpu.x[self.operands.rs2 as usize]),
+            );
         }
     }
 }
@@ -47,8 +48,8 @@ impl RISCVTrace for REM {
     fn trace(&self, cpu: &mut Cpu, trace: Option<&mut Vec<RV32IMCycle>>) {
         // RISCV spec: For REM, the sign of a nonzero result equals the sign of the dividend.
         // REM operands
-        let x = cpu.x[self.operands.rs1];
-        let y = cpu.x[self.operands.rs2];
+        let x = cpu.x[self.operands.rs1 as usize];
+        let y = cpu.x[self.operands.rs2 as usize];
 
         let (quotient, remainder) = match cpu.xlen {
             Xlen::Bit32 => {
@@ -98,10 +99,10 @@ impl RISCVTrace for REM {
 impl VirtualInstructionSequence for REM {
     fn virtual_sequence(&self) -> Vec<RV32IMInstruction> {
         // Virtual registers used in sequence
-        let v_0 = virtual_register_index(0) as usize;
-        let v_q = virtual_register_index(1) as usize;
-        let v_r = virtual_register_index(2) as usize;
-        let v_qy = virtual_register_index(3) as usize;
+        let v_0 = virtual_register_index(0);
+        let v_q = virtual_register_index(1);
+        let v_r = virtual_register_index(2);
+        let v_qy = virtual_register_index(3);
 
         let mut sequence = vec![];
