+++ tracer/src/instruction/slti.rs
@@ -17,11 +17,11 @@ declare_riscv_instr!(
 
 impl SLTI {
     fn exec(&self, cpu: &mut Cpu, _: &mut <SLTI as RISCVInstruction>::RAMAccess) {
-        cpu.x[self.operands.rd] = match cpu.x[self.operands.rs1] < normalize_imm(self.operands.imm)
-        {
-            true => 1,
-            false => 0,
-        };
+        cpu.x[self.operands.rd as usize] =
+            match cpu.x[self.operands.rs1 as usize] < normalize_imm(self.operands.imm) {
+                true => 1,
+                false => 0,
+            };
     }
 }
